requirements.txt
psutil
datetime
requests
pyqrcode
tasks
py-cord
mcstatus
pypng
